% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jip_comp.R
\name{jip_comp}
\alias{jip_comp}
\title{compute the parameters of jip test}
\usage{
jip_comp(df, use_PAM = FALSE)
}
\arguments{
\item{df}{data for jip test}

\item{use_PAM}{if TRUE use the PAM fluorescence signal, else use the continiuous signal}
}
\description{
only functions to calulate the parameters of jip test
}
\details{
provide core functions for jip_test, calculation based on Revisiting JIP-test: An educative review on concepts,
 assumptions, approximations, definitions and terminology
}
\examples{
\dontrun{
library(jiptest)
jip_comp(data)
}

}
