% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jip_dcplot.R
\name{jip_dcplot}
\alias{jip_dcplot}
\title{graphical preview of all the data curves in log axis (BLUE > 1.2.2)}
\usage{
jip_dcplot(file_dir)
}
\arguments{
\item{file_dir}{dataframe aquired by read_files.}
}
\description{
a quick way to preview the quality of measured data.
}
\details{
a quick preview the tendency of all measured data, based on ggplot2.
}
\examples{
\dontrun{
library(jiptest)
ojip <- read_files('./ojip1_2_2')
jip_dcplot(ojip)
}

}
