% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jip_dctest.R
\name{jip_dctest}
\alias{jip_dctest}
\title{apply jip test data analysis to the data measured by LI-6800 in batch}
\usage{
jip_dctest(file_dir)
}
\arguments{
\item{file_dir}{is the file directory only contains all the measured data saved as .xlsx.}
}
\description{
return jip test parameters from the induction data measured by
LI-6800.
}
\details{
the aim is to quickly calculate all the jip test parameters in batch, as one usually
take large number of data in one measurement.
}
\examples{
\dontrun{
library(jiptest)
jip_dctest('./ojip', 'd:/testdata/')
}

}
