% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jip_pca.R
\name{jip_pca}
\alias{jip_pca}
\title{return the prcomp}
\usage{
jip_pca(jip_data)
}
\arguments{
\item{jip_data}{data frame returned by \code{jip_test}}
}
\description{
{help to tidy the \code{jiptest} result to be ready for
PCA analysis }
}
