% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jip_test.R
\name{jip_test}
\alias{jip_test}
\title{apply jip test data analysis to the data measured by LI-6800}
\usage{
jip_test(df, use_PAM = FALSE)
}
\arguments{
\item{df}{dataframe that contains the induction data.}

\item{use_PAM}{if TRUE use the PAM fluorescence signal, else use the continiuous signal}
}
\description{
return jip test parameters from the induction data measured by
LI-6800.
}
\details{
the aim is to quickly calculate all the jip test parameters in batch, as one usually
take large number of data in one measurement.
}
\examples{
\dontrun{
library(jiptest)
jip_test('./ojip1_2_2', 'd:/testdata/')
}
}
