% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jiptest-method.R
\name{plot.jip}
\alias{plot.jip}
\title{plot jip test data measured by LI-6800}
\usage{
\method{plot}{jip}(
  df,
  alpha = 0.6,
  pch = 19,
  col,
  leg_bty = "n",
  leg_cex = 0.6,
  leg_point_cex = 0.9,
  legend_pos = "topleft",
  ylim = NULL,
  log = "x",
  xlab = "Time (Secs)",
  ylab = "Fluorescence signal",
  ...
)
}
\arguments{
\item{df}{files read by \code{read_files} or \code{read_dcfiles}}

\item{alpha}{the transparent degree of colors in the plot}

\item{pch}{inherited from plot.default}

\item{col}{inherited from plot.default}

\item{leg_bty}{bty in \code{legend}}

\item{leg_cex}{text size of legend}

\item{leg_point_cex}{point size of legend}

\item{legend_pos}{inherited from \code{legend} position}

\item{ylim}{Limits for the Y axis, if left blank estimated from data}

\item{log}{inherited from plot.default}

\item{xlab}{inherited from plot.default}

\item{ylab}{inherited from plot.default}

\item{...}{other parameters in \code{plot}}
}
\description{
{a quick way to preview the quality
of measured data.graphical preview of all the data
curves in log axis (BLUE > 1.2.2)}
}
