% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dcfiles.R
\name{read_dcfiles}
\alias{read_dcfiles}
\title{read all induction data in a file measured by LI-6800 (BLUE > 1.2.2)}
\usage{
read_dcfiles(file_dir)
}
\arguments{
\item{file_dir}{is the file directory only contains all the measured
data saved as .xlsx.}
}
\description{
use read_excel and a for loop to bombine all the data measured,
only contain modulation fluorescence for now.
}
\details{
the aim is to quickly import all data measured by LI-6800 to R for further use.
data are imported by read_excel from package readxl, by a for loop, the data are
all combine with rbind and we list the file name which the data come from.
}
\examples{
\dontrun{
library(jiptest)
read_dcfiles('./ojip')
}

}
