% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_induction.R
\name{read_induction}
\alias{read_induction}
\title{calculate the area under curve (BLUE > 1.2.2)}
\usage{
read_induction(path)
}
\arguments{
\item{path}{path of and excel file.}
}
\description{
read fluorescence files and keep the normal data
}
\details{
{read a single LI-6800 INDUCTION  EXCEL FILE}
}
\examples{
\dontrun{
library(jiptest)
read_induction('df.xlsx')
}
}
