% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_name.R
\name{sub_name}
\alias{sub_name}
\title{set the category of the data}
\usage{
sub_name(df, category)
}
\arguments{
\item{df}{data of a type data frame (contains the SOURCE column that specify which
excel files that the data have come from).}

\item{category}{a character verctor to group the data}
}
\description{
to separate the data related to the excel files. add a new column
of data that represent the treatment/replications etc.
}
\details{
use Soto's answer as
 https://stackoverflow.com/questions/4954507/calculate-the-area-under-a-curve
}
\examples{
\dontrun{
library(jiptest)
area_cal(df)
}
}
