% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jip_comp.R
\name{jip_comp}
\alias{jip_comp}
\title{compute the parameters of jip test}
\usage{
jip_comp(jip_data, normalized = FALSE)
}
\arguments{
\item{jip_data}{data for jip test}

\item{normalized}{use normalized data if TRUE}
}
\description{
only functions to calulate the parameters of jip test
}
\details{
provide core functions for jip_test, by changing the names of the data, it can also be used by other instruments.
}
\examples{
\dontrun{
library(jiptest)
jip_comp(data)
}

}
