% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jiptest-method.R
\name{plot.jip}
\alias{plot.jip}
\title{plot jip test data measured by LI-6800}
\usage{
\method{plot}{jip}(
  df,
  alpha = 0.6,
  pch = 19,
  col = NULL,
  leg_bty = "n",
  leg_cex = 0.6,
  leg_point_cex = 0.9,
  legend_pos = "topleft",
  xlim = c(5e-06, 1.1),
  ylim = NULL,
  log = "x",
  xlab = "Time (Secs)",
  ylab = "Fluorescence signal",
  xmark = c(expression(10^{     -5 }, 10^{     -4 }, 10^{     -3 }, 10^{     -2 }, 10^{
        -1 }, 10^0)),
  xat = c(1e-05, 1e-04, 0.001, 0.01, 0.1, 1),
  normalized = TRUE,
  add_leg = TRUE,
  ...
)
}
\arguments{
\item{df}{files read by \code{read_files} or \code{read_dcfiles}}

\item{alpha}{the transparent degree of colors in the plot}

\item{pch}{inherited from plot.default}

\item{col}{inherited from plot.default}

\item{leg_bty}{bty in \code{legend}}

\item{leg_cex}{text size of legend}

\item{leg_point_cex}{point size of legend}

\item{legend_pos}{inherited from \code{legend} position}

\item{xlim}{Limits for the Y axis, the default should be enough
for most cases}

\item{ylim}{Limits for the Y axis, if left blank estimated from data}

\item{log}{inherited from plot.default}

\item{xlab}{inherited from plot.default}

\item{ylab}{inherited from plot.default}

\item{xmark}{x tick marks values}

\item{xat}{at which the x axis tick marks are ploted}

\item{normalized}{whether use normorlized fluor signal}

\item{add_leg}{defaut is true, add lengend directly}

\item{...}{other parameters in \code{plot}}
}
\description{
{a quick way to preview the quality
of measured data.graphical preview of all the data
curves in log axis (BLUE > 1.2.2)}
}
